+++ src/main/java/com/iker/Lexly/Entity/Order.java
@@ -6,7 +6,7 @@ import java.util.List;
 public class Order {
     private List<Item> items;
 
-    private float amount;
+    private int amount;
     private String currency="DKK";
     private String reference;
 
@@ -19,11 +19,11 @@ public class Order {
         this.items = items;
     }
 
-    public float getAmount() {
+    public int getAmount() {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(int amount) {
         this.amount = amount;
     }
